requirements.txt
affine
attrs
beautifulsoup4
click
click-plugins
cligj
distlib
filelock
# Downgrade geospatial libs for Debian 11 (GDAL 3.2, PROJ 7.2, GEOS 3.9)
fiona<1.9
gdown
# GDAL is handled by manual install in Dockerfile
# Core processing libraries
geopandas<0.13
numpy<2.0.0
packaging
platformdirs
pluggy
pyproj
rasterio
requests
shapely<2.0
tqdm
urllib3
virtualenv
zstandard
scipy
# Cloud & API dependencies
fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-firestore
